requirements.txt
Django==1.11.13
django-appconf
django-ipware
geoip2
geopy
